requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
attrs==24.2.0
boto3==1.35.17
botocore==1.35.17
certifi==2024.7.4
charset-normalizer==3.3.2
et-xmlfile==1.1.0
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
idna==3.7
iniconfig==2.0.0
jmespath==1.0.1
multidict==6.1.0
numpy==2.1.0
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
playwright==1.46.0
pluggy==1.5.0
pyee==11.1.0
PySocks==1.7.1
pytest==8.3.2
pytest-base-url==2.1.0
pytest-playwright==0.5.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2024.1
requests==2.32.3
s3transfer==0.10.2
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
state==0.1.1.dev0
tabulate==0.9.0
text-unidecode==1.3
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
websocket-client==1.8.0
wsproto==1.2.0
yarl==1.11.1
